fixed gps for ios8
i fixed the damn thing!!!

+++ TrackMap/TrackMap/MapMenu/mapMenuViewController.m
@@ -48,13 +48,13 @@
 #pragma mark Table View Delegate Methods
 
 
-- (long)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     return self.menuItems.count;
 }
 
 
-- (double)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath;
+- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath;
 {
     return 60.0;
 }
